Budgeting!~

+++ ToTe/BudgetCategory.h
@@ -13,8 +13,8 @@
 @property(nonatomic)int budgetCategory_id;
 @property(nonatomic)int budget_id;
 @property(nonatomic)int category_id;
-@property(nonatomic)double category_spent;
-@property(nonatomic)double category_amount;
+@property(nonatomic)int category_spent;
+@property(nonatomic)int category_amount;
 @property(nonatomic)NSString *bcategory_name;
 @property(nonatomic)NSString *bcategory_image;
